pyproject.toml
[tool.poetry]
name = "techmed"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.115.3"}
uvicorn = "^0.32.0"
sqlmodel = "^0.0.22"
mysql-connector-python = "^9.1.0"
pyjwt = "^2.9.0"
passlib = {extras = ["argon2"], version = "^1.7.4"}
pyotp = "^2.9.0"
webauthn = {version = "^2.2.0", optional = true}
python-jose = "^3.3.0"


[tool.poetry.group.dev.dependencies]
alembic = "^1.13.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
